requirements.txt
torch
torchvision
torchaudio
transformers
peft
albumentations
torchmetrics
scipy
opencv-python
tqdm
numpy
matplotlib
pillow

